fix(lint): address ObsidianReviewBot findings
- drop `async` from `onload()` — no top-level await, the inner code-block processor is its own async callback
- remove redundant `as Uint8Array` assertion on `wasmBinary` — `init()`'s parameter is already typed
- lowercase "concept maps" / "timelines" in settings description (sentence case); acronyms VGL/IBIS/CLD kept as proper-noun usage

+++ src/vgraph-loader.ts
@@ -29,7 +29,7 @@ export async function initVithanco(): Promise<(vgl: string) => string> {
     // The WASM binary is embedded at build time (esbuild binary loader),
     // so init() never executes the import.meta.url fetch in Package/index.js.
     const { exports }: { exports: Exports } = await init({
-        module: wasmBinary as Uint8Array,
+        module: wasmBinary,
         getImports: () => ({}),
     });
